Added switch script from suckless.org, with tweak
Added a bit of a hack to hijack focusurgent.
Introduces a bit of a race condition, but it doesn't really matter-
just press Super+u until the right window is found.
The rationale for this is to stop windows from having the power to
hijack dwm for any reason and retain that power for the switch script.