[PATCH] implemented focuslast and focusjump.
focusjump is a wrapper for focus that saves the current selected client
to the monitor's "lastjump" pointer. It changes focus just as focus()
does, but does not change "lastjump" if the client to focus is NULL or
is already the last client to be focused.
By default, it is not used in any default dwm calls,
and is intended to be inserted by
the user in place of focus() calls that may warp client focus large distances.

focuslast jumps to the client in the "lastjump" pointer,
itself calling "focusjump", causing repeated calls to it to jump back
and forth between two windows.
By default, it is bound to MODKEY + ;.